Rename flow.py -> state.py
